# docs/swagger.yaml
...
info:
  contact: {}
paths:
...
  /verify/postcode:
    post:
      consumes:
      - application/json
      description: Checks if a postcode is valid returns 200 if successful
      parameters:
      - description: 'token: test'
        in: header
        name: Authorization
        required: true
        type: string
      - description: CR0 2HS
        in: query
        name: postcode
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ""
        "404":
          description: ""
      summary: Checks if a postcode is valid
swagger: "2.0"